Advanced topic upload
This file is an advanced reference not only to software developers, but is a reference to hardware designers too. I just concluded the translation from japanese to english, with help of yandex translator, to get translation of the pictures.
I attached here the original file I got, partially translated, as reference.
Thank you very much, HRA, the original writer of this remarkable document!